+++ BE/src/main/java/com/doit/dto/admin/GradeSubmissionRequest.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+
 
 @Data
 @Builder
@@ -20,7 +20,7 @@ public class GradeSubmissionRequest {
     @NotNull(message = "Band score is required")
     @DecimalMin(value = "0.0", message = "Band score must be at least 0.0")
     @DecimalMax(value = "9.0", message = "Band score must be at most 9.0")
-    private BigDecimal bandScore;
+    private Double bandScore;
     
     @NotBlank(message = "Feedback is required")
     private String feedback;
